kakao blindtest 2018 문제풀이소스 커밋 - 4번 완전 해결 (효율성테스트 2번 해결: 이유 (int x int < long 무한루프))

+++ JobDongSani/src/programmers/blindtest2018/Problem4Eff.java
@@ -21,7 +21,7 @@ public class Problem4Eff {
         Arrays.sort(sorted_food_times);
         int minIndex = 0;
         int minCapSum = 0;
-        while(minCapFood * foodNums <= k){
+        while((long) minCapFood * (long) foodNums <= k){
             int nums = 0;
             minCapSum = sorted_food_times[minIndex];
             for (int i = minIndex; i < sorted_food_times.length; i++) {
